cleanup(core): move tools utils to separate file
